+++ src/Options.js
@@ -11,9 +11,10 @@ export default class Options {
     strictDelete: true,
     alwaysShowDefaultError: false,
     defaultGenericError: null,
-    loadingMessage: null,
     allowRequired: false,
     autoParseNumbers: false,
+    validationDebounceWait: 250,
+    validationDebounceOptions: { leading: true },
   };
 
   get(key = null) {